# src/config/swagger.json
{
  "openapi": "3.0.3",
  "servers": [{ "url": "" }],
  "info": {
    "title": "Monix Documentation API",
    "version": "1.0",
    "description": "La documentation du backend de monix (le logiciel de gestion de crédits du club nix)"
  },
...
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Connecte l'utilisateur",
        "tags": ["Authentification"],
        "requestBody": {
          "description": "L'utilisateur a créer'",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "201": {
            "description": "L'utilisateur a ete cree",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/codeLogin": {
      "post": {
        "summary": "Connecte l'utilisateur en utilisant son code personnel",
        "tags": ["Authentification"],
        "requestBody": {
          "description": "Le code d'authentification",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/codeLoginBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "201": {
            "description": "L'utilisateur a ete cree",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgottenPassword": {
      "post": {
        "summary": "Envoie à l'utilisateur un lien pour reset son mot de passe par mail",
        "tags": ["Authentification"],
        "requestBody": {
          "description": "Utilisateur a qui envoyer un mot de passe",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/forgottenPasswordBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "201": {
            "description": "Le mail a bien été envoyé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/loginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/resetPassword": {
      "post": {
        "summary": "Change the password using the code received by email",
        "tags": ["Authentification"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/resetPasswordBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "Le mot de passe a bien été changé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiMessage"
                }
              }
            }
          }
        }
      }
    },
...
    "/balance/buy": {
      "post": {
        "summary": "Buy a product for the authentified user",
        "tags": ["Balance"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": { "type": "number" },
                  "amount": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "Le produit a bien été acheté",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGet"
                }
              }
            }
          }
        }
      }
    },
    "/balance/recharge": {
      "post": {
        "summary": "Endpoint permettant à l'utilisateur de recharger son portefeuille",
        "tags": ["Balance"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "Le compte a bien été rechargé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGet"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "apiError": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "apiMessage": {
        "type": "object",
        "required": ["success", "message"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "loginBody": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "forgottenPasswordBody": {
        "type": "object",
        "required": ["username"],
        "properties": {
          "username": {
            "type": "string"
          }
        }
      },
      "resetPasswordBody": {
        "type": "object",
        "required": ["code", "newPassword", "passwordConfirmation"],
        "properties": {
          "code": { "type": "string" },
          "newPassword": { "type": "string" },
          "passwordConfirmation": { "type": "string" }
        }
      },
      "codeLoginBody": {
        "type": "object",
        "required": ["code"],
        "properties": {
          "code": {
            "type": "string"
          }
        }
      },
      "loginResponse": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "required": ["userId", "email", "token"],
            "properties": {
              "userId": { "type": "number" },
              "email": { "type": "string" },
              "token": { "type": "string" }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "balance",
          "avatar",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "type": "number" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "balance": { "type": "number" },
          "avatar": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        }
      },
...
      "UserGet": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/User" }
        }
      },
...
    }
  }
}